ADD: Register NBT transfer recipe serializers
Integrates custom recipe serializers with Forge registration
system and mod initialization.

Changes:
- LovelyRecipes.java: New registry class with DeferredRegister
  for LOVELY_SPAWN and LOVELY_SPAWN_DYE serializers
- LovelyLegacy.java: Registers recipe serializers on mod bus
- InternalItems.java: Minor formatting/organizational updates

Uses Forge's DeferredRegister pattern for proper initialization
timing during mod loading.

Related: Sprint 01 - NBT Recipe System Implementation

+++ sources/legacy/llovelyr-1.20.1/Forge/src/main/java/net/msymbios/llovelyr/common/item/InternalItems.java
@@ -68,9 +68,9 @@ public class InternalItems {
     protected static void registerModel(Item item, ResourceLocation tag, String key) {
         ItemProperties.register(item, tag, (stack, world, entity, seed) -> {
             if (stack.hasTag() && stack.getTag().contains(key)) {
-                return (float) stack.getTag().getInt(key);
+                return stack.getTag().getInt(key);
             }
-            return 16.0f;
+            return 16;
         });
     } // registerModel()
